requirements.txt
google-generativeai
django
pillow
torch
torchvision
opencv-python
numpy
joblib
scikit-learn
matplotlib
